requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-generativeai
pytrends
python-dotenv
pandas
requests
beautifulsoup4
lxml
